requirements.txt
python-dotenv==1.0.0
velebit-useful-logs==1.0.3
pyodbc==5.0.0